+++ src/Todo.psm1
@@ -1,17 +1,17 @@
 # import classes
-if (Test-Path "$PSScriptRoot\classes\classes.psd1") {
-    $ClassLoadOrder = Import-PowerShellDataFile -Path "$PSScriptRoot\Classes\classes.psd1" -ErrorAction SilentlyContinue
+if (Test-Path "${PSScriptRoot}\classes\classes.psd1") {
+    $ClassLoadOrder = Import-PowerShellDataFile -Path "${PSScriptRoot}\classes\classes.psd1" -ErrorAction SilentlyContinue
 }
 
-foreach ($class in $ClassLoadOrder.order) {
-    $Path = '{0}\classes\{1}.ps1' -f $PSScriptRoot, $class
+foreach ($Class in $ClassLoadOrder.Order) {
+    $Path = '{0}\classes\{1}.ps1' -f $PSScriptRoot, $Class
     if (Test-Path $Path) {
         . $Path
     }
 }
 
 # dot source ps1 files
-$Public  = @( Get-ChildItem -Path $PSScriptRoot\Public\*.ps1 -ErrorAction SilentlyContinue )
+$Public  = @( Get-ChildItem -Path "${PSScriptRoot}\public\*.ps1" -ErrorAction SilentlyContinue )
 
 foreach ($Import in @($Public)) {
     try {
@@ -23,4 +23,4 @@ foreach ($Import in @($Public)) {
     }
 }
 
-Export-ModuleMember -Function $Public.Basename
+Export-ModuleMember -Function $Public.Basename -Alias *