+++ CHANGELOG.md
@@ -2,4 +2,11 @@
 
 ## Version 1.0.0 (2022-02-20)
 
-TODO
+Initial release of the `Todo` module. Implements the following Cmdlets:
+
+- `New-TodoList`
+- `Get-TodoList`
+- `Remove-TodoList`
+- `New-Task`
+- `Update-Task`
+- `Remove-Task`